代码重构：

+++ platformservlet/src/main/java/com/dao/ActivityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.model.Activity;
 
-public interface ActivityRepository extends JpaRepository<Activity,Long>{
+public interface ActivityRepository extends JpaRepository<Activity,Integer>{
 
 }
